refactor: 경매 상세 api auction.ts로 변경
